refactor: move distribution behavior to package
Readers are more likely to look for this behavior class in the
distribution package than in the common package. I noticed that it
wasn't there when I was looking for it myself. I think it makes sense to
keep the functionality close together, so I've placed it next to the
CommandRedistributor and the CommandDistributionAcknowledgeProcessor.
